+++ src/libAtomVM/debug.c
@@ -85,7 +85,7 @@ COLD_FUNC void debug_dump_context(Context *ctx)
 
 COLD_FUNC void debug_dump_heap(Context *ctx)
 {
-    debug_dump_memory(ctx, ctx->heap.heap_start, ctx->heap.heap_ptr, "heap");
+    debug_dump_memory(ctx, memory_heap_start(&ctx->heap), ctx->heap.heap_ptr, "heap");
 }
 
 COLD_FUNC void debug_dump_stack(Context *ctx)