+++ src/libAtomVM/erl_nif_priv.h
@@ -60,9 +60,9 @@ static inline void erl_nif_env_partial_init_from_globalcontext(ErlNifEnv *env, G
 {
     env->global = global;
     env->heap.root = NULL;
-    env->heap.heap_start = NULL;
     env->heap.heap_ptr = NULL;
     env->heap.heap_end = NULL;
+    env->heap.fragments_words = 0;
     env->stack_pointer = NULL;
     env->x[0] = term_nil();
     env->x[1] = term_nil();
@@ -73,9 +73,9 @@ static inline void erl_nif_env_partial_init_from_resource(ErlNifEnv *env, void *
     struct RefcBinary *refc = refc_binary_from_data(resource);
     env->global = refc->resource_type->global;
     env->heap.root = NULL;
-    env->heap.heap_start = NULL;
     env->heap.heap_ptr = NULL;
     env->heap.heap_end = NULL;
+    env->heap.fragments_words = 0;
     env->stack_pointer = NULL;
     env->x[0] = term_nil();
     env->x[1] = term_nil();